feat: drop string support in message functions
BREAKING CHANGE: Previously, `.getMessage` and `.compatibleMessage`
accepted both `Error` instances and strings. This commit removes
support for strings. This is being done in preparation for the
upcoming addition of a `.checkError` method that'll require the
thrown object to be an `Error` instance. With that requirement in
place, it's no longer necessary to support strings in `.getMessage`
and `.compatibleMessage`.